requirements.txt
# Flask ve ilgili uzantılar
flask==2.2.3
flask-restful==0.3.9
flask-cors==3.0.10
flask-jwt-extended==4.4.4
marshmallow==3.19.0
python-dotenv==1.0.0
werkzeug==2.2.3
gunicorn==20.1.0

# AWS ve DynamoDB bağlantıları
boto3==1.26.84
pynamodb==5.4.1

# Yardımcı kütüphaneler
bcrypt==4.0.1
pillow==9.4.0
python-magic==0.4.27
uuid==1.30
email-validator==1.3.1
requests==2.28.2
structlog==23.1.0
python-dateutil==2.8.2

# Bağlantı havuzu
aiobotocore==2.5.0

# Asenkron operasyonlar için
celery==5.2.7
redis==4.5.1

# Güvenlik
pyjwt==2.6.0